#14841 Move browser stuff to browser folder
