+++ src/vs/workbench/parts/extensions/browser/media/extensionEditor.css
