+++ src/vs/workbench/parts/extensions/browser/media/extensionsWidgets.css
