+++ src/vs/workbench/parts/extensions/common/media/defaultIcon.png
