+++ src/vs/workbench/parts/extensions/electron-browser/extensions.contribution.ts
@@ -12,7 +12,7 @@ import { registerSingleton } from 'vs/platform/instantiation/common/extensions';
 import { IExtensionGalleryService, IExtensionTipsService, ExtensionsLabel, ExtensionsChannelId } from 'vs/platform/extensionManagement/common/extensionManagement';
 import { ExtensionGalleryService } from 'vs/platform/extensionManagement/node/extensionGalleryService';
 import { IWorkbenchActionRegistry, Extensions as WorkbenchActionExtensions } from 'vs/workbench/common/actionRegistry';
-import { ExtensionTipsService } from 'vs/workbench/parts/extensions/electron-browser/extensionTipsService';
+import { ExtensionTipsService } from 'vs/workbench/parts/extensions/browser/extensionTipsService';
 import { IWorkbenchContributionsRegistry, Extensions as WorkbenchExtensions } from 'vs/workbench/common/contributions';
 import { IOutputChannelRegistry, Extensions as OutputExtensions } from 'vs/workbench/parts/output/common/output';
 import { EditorDescriptor } from 'vs/workbench/browser/parts/editor/baseEditor';
@@ -28,8 +28,8 @@ import {
 import { OpenExtensionsFolderAction, InstallVSIXAction } from 'vs/workbench/parts/extensions/electron-browser/extensionsActions';
 import { ExtensionsInput } from 'vs/workbench/parts/extensions/common/extensionsInput';
 import { ViewletRegistry, Extensions as ViewletExtensions, ViewletDescriptor } from 'vs/workbench/browser/viewlet';
-import { ExtensionEditor } from './extensionEditor';
-import { StatusUpdater } from './extensionsViewlet';
+import { ExtensionEditor } from 'vs/workbench/parts/extensions/browser/extensionEditor';
+import { StatusUpdater } from 'vs/workbench/parts/extensions/electron-browser/extensionsViewlet';
 import { IQuickOpenRegistry, Extensions, QuickOpenHandlerDescriptor } from 'vs/workbench/browser/quickopen';
 import { IConfigurationRegistry, Extensions as ConfigurationExtensions } from 'vs/platform/configuration/common/configurationRegistry';
 import jsonContributionRegistry = require('vs/platform/jsonschemas/common/jsonContributionRegistry');
@@ -49,7 +49,7 @@ Registry.as<IOutputChannelRegistry>(OutputExtensions.OutputChannels)
 // Quickopen
 Registry.as<IQuickOpenRegistry>(Extensions.Quickopen).registerQuickOpenHandler(
 	new QuickOpenHandlerDescriptor(
-		'vs/workbench/parts/extensions/electron-browser/extensionsQuickOpen',
+		'vs/workbench/parts/extensions/browser/extensionsQuickOpen',
 		'ExtensionsHandler',
 		'ext ',
 		localize('extensionsCommands', "Manage Extensions"),
@@ -59,7 +59,7 @@ Registry.as<IQuickOpenRegistry>(Extensions.Quickopen).registerQuickOpenHandler(
 
 Registry.as<IQuickOpenRegistry>(Extensions.Quickopen).registerQuickOpenHandler(
 	new QuickOpenHandlerDescriptor(
-		'vs/workbench/parts/extensions/electron-browser/extensionsQuickOpen',
+		'vs/workbench/parts/extensions/browser/extensionsQuickOpen',
 		'GalleryExtensionsHandler',
 		'ext install ',
 		localize('galleryExtensionsCommands', "Install Gallery Extensions"),
@@ -71,7 +71,7 @@ Registry.as<IQuickOpenRegistry>(Extensions.Quickopen).registerQuickOpenHandler(
 const editorDescriptor = new EditorDescriptor(
 	ExtensionEditor.ID,
 	localize('extension', "Extension"),
-	'vs/workbench/parts/extensions/electron-browser/extensionEditor',
+	'vs/workbench/parts/extensions/browser/extensionEditor',
 	'ExtensionEditor'
 );
