+++ src/vs/workbench/workbench.main.ts
@@ -54,7 +54,7 @@ import 'vs/workbench/parts/markers/browser/markersPanel'; // can be packaged sep
 import 'vs/workbench/parts/html/browser/html.contribution';
 
 import 'vs/workbench/parts/extensions/electron-browser/extensions.contribution';
-import 'vs/workbench/parts/extensions/electron-browser/extensionsQuickOpen';
+import 'vs/workbench/parts/extensions/browser/extensionsQuickOpen';
 import 'vs/workbench/parts/extensions/electron-browser/extensionsViewlet'; // can be packaged separately
 
 import 'vs/workbench/parts/output/browser/output.contribution';
